@@ -4,10 +4,6 @@ package ru.proshik.applepricebot.service.provider;
 import org.springframework.stereotype.Component;
 import ru.proshik.applepricebot.repository.model.Provider;
 import ru.proshik.applepricebot.repository.model.ProviderType;
-import ru.proshik.applepricebot.service.provider.aj.AjScreening;
-import ru.proshik.applepricebot.service.provider.citilink.CitilinkSpbScreening;
-import ru.proshik.applepricebot.service.provider.gsmstore.GsmStoreScreening;
-import ru.proshik.applepricebot.service.provider.istorespb.IStoreSpbScreening;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -15,22 +11,22 @@ import java.util.Map;
 @Component
 public class ProviderResolver {
 
-    private Map<ProviderType, Screening> PROVIDERS = new HashMap<>();
+    private Map<ProviderType, ScreeningProvider> PROVIDERS = new HashMap<>();
 
     public ProviderResolver() {
-        PROVIDERS.put(ProviderType.AJ, new AjScreening());
-        PROVIDERS.put(ProviderType.GSM_STORE, new GsmStoreScreening());
-        PROVIDERS.put(ProviderType.ISTORE_SBP, new IStoreSpbScreening());
-        PROVIDERS.put(ProviderType.CITI_LINK, new CitilinkSpbScreening());
+        PROVIDERS.put(ProviderType.AJ, new AjScreeningProvider());
+        PROVIDERS.put(ProviderType.GSM_STORE, new GsmStoreScreeningProvider());
+        PROVIDERS.put(ProviderType.ISTORE_SBP, new IStoreSpbScreeningProvider());
+        PROVIDERS.put(ProviderType.CITI_LINK, new CitilinkSpbScreeningProvider());
     }
 
-    public Screening resolve(Provider provider) {
-        Screening screening = PROVIDERS.get(provider.getType());
-        if (screening == null) {
-            throw new IllegalArgumentException("Not found screening module for type=" + provider.getType());
+    public ScreeningProvider resolve(Provider provider) {
+        ScreeningProvider screeningProvider = PROVIDERS.get(provider.getType());
+        if (screeningProvider == null) {
+            throw new IllegalArgumentException("Not found screeningProvider module for type=" + provider.getType());
         }
 
-        return screening;
+        return screeningProvider;
     }
 
 }
